@@ -11,7 +11,7 @@
     "@types/eslint": "^9.6.0",
     "typescript-eslint": "^8.0.0",
     "prettier": "^3.3.3",
-    "eslint": "^9.7.0",
+    "eslint": "^9.14.0",
     "eslint-config-prettier": "^9.1.0",
     "prettier-plugin-svelte": "^3.2.6"
   }
