@@ -1,26 +1,30 @@
 import { BackendMethod, remult } from 'remult';
-import { lucia } from '../../../src/lib/auth/lucia';
-import { authRepo } from '../..';
+import { authRepo } from '../../index';
 
 export class UserController {
-	@BackendMethod({ allowed: true })
-	static async signOut() {
-		if (remult.user?.session.id) {
-			await lucia.invalidateSession(remult.user.session.id);
-			remult.user = undefined;
-		}
-		return { success: true };
-	}
+  @BackendMethod({ allowed: true })
+  static async signOut() {
+    if (remult.user?.session.id) {
+      await authRepo.session.delete(remult.user.session.id);
+      remult.user = undefined;
+    }
+    return { success: true };
+  }
 
-	@BackendMethod({ allowed: true })
-	static async signInDemo(name: string) {
-		let user = await authRepo.user.findFirst({ username: name });
+  @BackendMethod({ allowed: true })
+  static async signInDemo(name: string) {
+    let user = await authRepo.user.findFirst({ username: name });
 
-		if (!user) {
-			user = authRepo.user.create();
-			user.username = name;
-			user.roles = ['Admin'];
-			await authRepo.user.save(user);
-		}
-	}
+    if (!user) {
+      user = authRepo.user.create();
+      user.username = name;
+      user.roles = ['Admin'];
+      await authRepo.user.save(user);
+    }
+  }
+
+  @BackendMethod({ allowed: true })
+  static async currentUser() {
+    return remult.user;
+  }
 }
