@@ -1,25 +1,25 @@
 import { Entity, Fields, Validators, Allow } from 'remult';
 
 @Entity<Product>('products', {
-	allowApiCrud: Allow.authenticated,
-	dbName: 'products'
+  allowApiCrud: Allow.authenticated,
+  dbName: 'products'
 })
 export class Product {
-	@Fields.cuid()
-	id!: string;
+  @Fields.cuid()
+  id!: string;
 
-	@Fields.string({ validate: [Validators.required] })
-	name!: string;
+  @Fields.string({ validate: [Validators.required] })
+  name!: string;
 
-	@Fields.number({ validate: [Validators.required] })
-	price!: number;
+  @Fields.number({ validate: [Validators.required] })
+  price!: number;
 
-	@Fields.string()
-	description?: string;
+  @Fields.string()
+  description?: string;
 
-	@Fields.createdAt()
-	createdAt!: Date;
+  @Fields.createdAt()
+  createdAt!: Date;
 
-	@Fields.updatedAt()
-	updatedAt?: Date;
+  @Fields.updatedAt()
+  updatedAt?: Date;
 }
